Migrate the processCommand test to jasmine
