+++ test/unit.html
@@ -37,7 +37,6 @@
   <script src="isArray_test_old.js"></script>
   <script src="isPlainObject_test_old.js"></script>
   <script src="merge_test_old.js"></script>
-  <script src="processCommand_test_old.js"></script>
   <script src="processStates_test_old.js"></script>
   <script src="type_test_old.js"></script>
